The use of to_string/1 for string
the `to_string/1` function appears to have use for converting a datatype back to a string.  However, it is also used to convert the default value as read to a string to add to the default section of the file.

Most to_string/1 clauses handle this by checking if the input is a list - and don't do anything with a list ... but a list will be a list, and a string!

Changed this so that it cannot be a list of integers - then we assume a list of all integers is in fact a string which doesn't need conversion, and not a list that needs conversion